requirements.txt
twikit
pytz
requests
redis
schedule
telethon
python-telegram-bot
apscheduler
pysocks
celery
aiohttp